Format code
stree write "lib/**/*.rb"

+++ lib/vandelay/assets/generators/copy_file.rb
@@ -9,15 +9,15 @@ module Vandelay
   module Assets
     module Generators
       class CopyFile < Base
-        sig {params(source_path: String).void}
+        sig { params(source_path: String).void }
         def initialize(source_path:)
           @source_path = source_path
         end
 
-        sig {override.params(path: String).void}
+        sig { override.params(path: String).void }
         def call(path)
-          return if unless File.exist?(path)
-            FileUtils.copy_file(@source_path, path)
+          if FileUtils.copy_file(@source_path, path) unless File.exist?(path)
+            return
           end
         end
       end
